Split cold_snap.rs into cold_snap/ subdirectory (1168 -> ~400 lines each)
Break the monolithic cold_snap.rs into focused submodules:
- types.rs (165 lines): constants, ColdSnapTier, ColdSnapState, ColdSnapEvent
- detection.rs (300 lines): cold day detection, tier classification, effect modifiers
- pipe_damage.rs (340 lines): PRNG, pipe burst probability, water service calculations
- systems.rs (402 lines): Bevy system, plugin, and system integration tests
- mod.rs (21 lines): re-exports preserving the public API

Part of #1389
